requirements.txt
ccxt==4.2.68
numpy==1.26.4
pandas==2.2.1
python_binance==1.0.19
TA_Lib==0.4.28
